+++ util/annotation.py
@@ -20,6 +20,8 @@ NTP_OFFSET_SECS = 2208988800
 #
 # ???         -       - ???     - Initial Commit
 # Sep 22 2017 - 11444 - jelkins - Prevent None endDT from crashing code
+# Oct 19 2017 - 12784 - sgfoote - Added parameters to AnnotationRecord and created a dump_json method to facilitate 
+#                                 writing to files
 #