+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Stream Engine
 
+# Next Release
+
+Issue #11543 - Add annotations to NetCDF and CSV requests
+
 # Release 1.3.8 2017-09-29
 
 Issue #11444 - Prevent crash on open ended annotations